+++ src/extended-iterable.ts
@@ -1,13 +1,13 @@
-import { BaseIterator } from './iterators/base-iterator.js';
 import { ConcatIterator } from './iterators/concat-iterator.js';
 import { DropIterator } from './iterators/drop-iterator.js';
 import { FilterIterator } from './iterators/filter-iterator.js';
-import { TakeIterator } from './iterators/take-iterator.js';
+import { FlatMapIterator } from './iterators/flatmap-iterator.js';
+import { MapErrorIterator } from './iterators/maperror-iterator.js';
+import { MapIterator } from './iterators/map-iterator.js';
 import { resolveIterator } from './iterators/resolve-iterator.js';
 import { SliceIterator } from './iterators/slice-iterator.js';
+import { TakeIterator } from './iterators/take-iterator.js';
 import type { IterableLike } from './types.js';
-import { MapIterator } from './iterators/map-iterator.js';
-import { FlatMapIterator } from './iterators/flatmap-iterator.js';
 
 /**
  * An iterable that provides a rich set of methods for working with ranges of
@@ -619,97 +619,10 @@ export class ExtendedIterable<T> {
 	 * ```
 	 */
 	mapError(catchCallback?: (error: Error | unknown) => Error | unknown | Promise<Error | unknown>): ExtendedIterable<T | Error> {
-		if (catchCallback && typeof catchCallback !== 'function') {
-			throw new TypeError('Callback is not a function');
-		}
-
 		return new ExtendedIterable(
-			new ExtendedIterable.ErrorMappingIterator(this.#iterator, catchCallback)
+			new MapErrorIterator(this.#iterator, catchCallback)
 		);
 	}
-
-	static ErrorMappingIterator = class ErrorMappingIterator<T> extends BaseIterator<T> {
-		#catchCallback?: (error: Error | unknown) => Error | unknown | Promise<Error | unknown>;
-
-		constructor(
-			iterator: Iterator<T> | AsyncIterator<T>,
-			catchCallback?: (error: Error | unknown) => Error | unknown | Promise<Error | unknown>
-		) {
-			super(iterator);
-			this.#catchCallback = catchCallback;
-		}
-
-		next(): IteratorResult<T | Error> | Promise<IteratorResult<T | Error>> | any {
-			try {
-				const result = super.next();
-
-				// async handling
-				if (result instanceof Promise) {
-					return this.#asyncNext(result);
-				}
-
-				// sync handling
-				if (result.done) {
-					return result;
-				}
-
-				return {
-					done: false,
-					value: result.value
-				};
-			} catch (error: unknown) {
-				// handle sync errors - return error as value and continue on next call
-				if (this.#catchCallback) {
-					const err = this.#catchCallback(error);
-					// if catchCallback returns a promise, switch to async handling
-					if (err instanceof Promise) {
-						return err.then(resolvedErr => ({
-							done: false,
-							value: resolvedErr
-						}));
-					}
-					return {
-						done: false,
-						value: err
-					};
-				}
-				return {
-					done: false,
-					value: error
-				};
-			}
-		}
-
-		async #asyncNext(result: Promise<IteratorResult<T>>): Promise<IteratorResult<T | Error>> {
-			try {
-				const currentResult = await result;
-
-				if (currentResult.done) {
-					return currentResult as IteratorResult<T | Error>;
-				}
-
-				return {
-					done: false,
-					value: currentResult.value
-				};
-			} catch (error) {
-				// handle async errors - return error as value
-				if (this.#catchCallback) {
-					const err = this.#catchCallback(error);
-					// await the result if it's a promise
-					const resolvedErr = err instanceof Promise ? await err : err;
-					return {
-						done: false,
-						value: resolvedErr as any
-					};
-				}
-				return {
-					done: false,
-					value: error as any
-				};
-			}
-		}
-	};
 
 	/**
 	 * Reduces the iterable to a single value.
